+++ enterprise/internal/database/BUILD.bazel
@@ -4,7 +4,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "database",
     srcs = [
-        "authz.go",
         "database.go",
         "external_services.go",
         "mocks_temp.go",
@@ -12,20 +11,17 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/database",
     visibility = ["//enterprise:__subpackages__"],
     deps = [
-        "//cmd/frontend/globals",
         "//enterprise/internal/authz/bitbucketserver",
         "//enterprise/internal/authz/github",
         "//enterprise/internal/authz/gitlab",
         "//enterprise/internal/authz/perforce",
         "//enterprise/internal/github_apps/store",
-        "//internal/authz",
         "//internal/database",
         "//internal/database/basestore",
         "//internal/database/dbconn",
         "//internal/database/dbutil",
         "//internal/encryption",
         "//internal/types",
-        "//lib/errors",
         "//schema",
         "@com_github_sourcegraph_log//:log",
     ],
@@ -36,7 +32,6 @@ go_test(
     size = "large",
     timeout = "long",
     srcs = [
-        "authz_test.go",
         "db_test.go",
         "external_services_test.go",
         "main_test.go",
@@ -47,22 +42,14 @@ go_test(
         "requires-network",
     ],
     deps = [
-        "//cmd/frontend/globals",
-        "//internal/api",
-        "//internal/authz",
         "//internal/conf",
         "//internal/database",
-        "//internal/database/dbtest",
         "//internal/extsvc",
         "//internal/licensing",
-        "//internal/types",
         "//lib/errors",
         "//schema",
         "@com_github_google_go_cmp//cmp",
         "@com_github_kylelemons_godebug//pretty",
         "@com_github_sourcegraph_log//logtest",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-        "@org_golang_x_exp//slices",
     ],
 )
